split modification nodes
