+++ arangod/Aql/ExecutionPlan.cpp
@@ -38,15 +38,19 @@
 #include "Aql/ExecutionNode/FilterNode.h"
 #include "Aql/ExecutionNode/IResearchViewNode.h"
 #include "Aql/ExecutionNode/IndexNode.h"
+#include "Aql/ExecutionNode/InsertNode.h"
 #include "Aql/ExecutionNode/LimitNode.h"
-#include "Aql/ExecutionNode/ModificationNodes.h"
 #include "Aql/ExecutionNode/NoResultsNode.h"
+#include "Aql/ExecutionNode/RemoveNode.h"
+#include "Aql/ExecutionNode/ReplaceNode.h"
 #include "Aql/ExecutionNode/ReturnNode.h"
 #include "Aql/ExecutionNode/ShortestPathNode.h"
 #include "Aql/ExecutionNode/SingletonNode.h"
 #include "Aql/ExecutionNode/SortNode.h"
 #include "Aql/ExecutionNode/SubqueryNode.h"
 #include "Aql/ExecutionNode/TraversalNode.h"
+#include "Aql/ExecutionNode/UpdateNode.h"
+#include "Aql/ExecutionNode/UpsertNode.h"
 #include "Aql/ExecutionNode/WindowNode.h"
 #include "Aql/Expression.h"
 #include "Aql/Function.h"
