+++ arangod/Aql/OptimizerRules.cpp
@@ -45,11 +45,14 @@
 #include "Aql/ExecutionNode/GatherNode.h"
 #include "Aql/ExecutionNode/IResearchViewNode.h"
 #include "Aql/ExecutionNode/IndexNode.h"
+#include "Aql/ExecutionNode/InsertNode.h"
 #include "Aql/ExecutionNode/JoinNode.h"
 #include "Aql/ExecutionNode/LimitNode.h"
 #include "Aql/ExecutionNode/MaterializeRocksDBNode.h"
-#include "Aql/ExecutionNode/ModificationNodes.h"
+#include "Aql/ExecutionNode/ModificationNode.h"
 #include "Aql/ExecutionNode/RemoteNode.h"
+#include "Aql/ExecutionNode/RemoveNode.h"
+#include "Aql/ExecutionNode/ReplaceNode.h"
 #include "Aql/ExecutionNode/ReturnNode.h"
 #include "Aql/ExecutionNode/ScatterNode.h"
 #include "Aql/ExecutionNode/ShortestPathNode.h"
@@ -58,6 +61,8 @@
 #include "Aql/ExecutionNode/SubqueryNode.h"
 #include "Aql/ExecutionNode/SubqueryStartExecutionNode.h"
 #include "Aql/ExecutionNode/TraversalNode.h"
+#include "Aql/ExecutionNode/UpdateNode.h"
+#include "Aql/ExecutionNode/UpsertNode.h"
 #include "Aql/ExecutionNode/WindowNode.h"
 #include "Aql/ExecutionPlan.h"
 #include "Aql/Expression.h"
