+++ arangod/Aql/OptimizerRulesCluster.cpp
@@ -30,9 +30,12 @@
 #include "Aql/ExecutionNode/EnumerateListNode.h"
 #include "Aql/ExecutionNode/ExecutionNode.h"
 #include "Aql/ExecutionNode/IndexNode.h"
-#include "Aql/ExecutionNode/ModificationNodes.h"
+#include "Aql/ExecutionNode/InsertNode.h"
+#include "Aql/ExecutionNode/ModificationNode.h"
 #include "Aql/ExecutionNode/MultipleRemoteModificationNode.h"
+#include "Aql/ExecutionNode/RemoveNode.h"
 #include "Aql/ExecutionNode/SingleRemoteOperationNode.h"
+#include "Aql/ExecutionNode/UpdateReplaceNode.h"
 #include "Aql/ExecutionPlan.h"
 #include "Aql/Expression.h"
 #include "Aql/Optimizer.h"
