+++ arangod/Aql/OptimizerUtils.cpp
@@ -33,9 +33,10 @@
 #include "Aql/ExecutionNode/GatherNode.h"
 #include "Aql/ExecutionNode/IResearchViewNode.h"
 #include "Aql/ExecutionNode/IndexNode.h"
-#include "Aql/ExecutionNode/ModificationNodes.h"
+#include "Aql/ExecutionNode/RemoveNode.h"
 #include "Aql/ExecutionNode/SubqueryNode.h"
 #include "Aql/ExecutionNode/TraversalNode.h"
+#include "Aql/ExecutionNode/UpdateReplaceNode.h"
 #include "Aql/ExecutionPlan.h"
 #include "Aql/Expression.h"
 #include "Aql/NonConstExpressionContainer.h"
